# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Multi-Auth Realtime API",
    "version": "1.0.0",
    "description": "Demonstrates Non-Auth, Basic, Digest, OAuth 1.0, OAuth 1.0a, OAuth 2.0 with realtime (WS, SSE, Socket.IO) - Single BaseUrl for all auth types"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Single BaseUrl for all auth types - Dedicated Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "digestAuth": {
        "type": "http",
        "scheme": "digest"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth1": {
        "type": "oauth2",
        "flows": {}
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:3000/api/oauth2/authorize",
            "tokenUrl": "http://localhost:3000/api/oauth2/token",
            "scopes": {
              "read": "read",
              "write": "write"
            }
          },
          "clientCredentials": {
            "tokenUrl": "http://localhost:3000/api/oauth2/token",
            "scopes": {
              "read": "read"
            }
          },
          "password": {
            "tokenUrl": "http://localhost:3000/api/oauth2/token",
            "scopes": {
              "read": "read",
              "write": "write"
            }
...
  },
  "paths": {
...
    "/api/oauth1/data": {
      "get": {
        "tags": [
          "OAuth 1.0"
        ],
        "summary": "OAuth 1.0 data",
        "security": [
          {
            "oauth1": []
          }
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "OAuth header with HMAC-SHA1 signature",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/oauth2/authorize": {
      "get": {
        "tags": [
          "OAuth 2.0"
        ],
        "summary": "Authorization endpoint",
        "parameters": [
          {
            "name": "response_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "code"
              ]
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "code"
          }
        }
      }
    },
    "/api/oauth2/token": {
      "post": {
        "tags": [
          "OAuth 2.0"
        ],
        "summary": "Token endpoint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "enum": [
                      "authorization_code",
                      "client_credentials",
                      "password",
                      "refresh_token"
                    ]
                  },
                  "code": {
                    "type": "string"
                  },
                  "client_id": {
                    "type": "string"
                  },
                  "client_secret": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "refresh_token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "token"
          }
        }
      }
    },
    "/api/oauth2/data": {
      "get": {
        "tags": [
          "OAuth 2.0"
        ],
        "summary": "OAuth2 protected data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/oauth2/introspect": {
      "post": {
        "tags": [
          "OAuth 2.0"
        ],
        "summary": "Introspect token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/realtime/sse/oauth2": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "summary": "SSE OAuth2 stream",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "stream"
          }
        }
      }
    },
    "/api/realtime/sse/basic": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "summary": "SSE Basic stream",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "stream"
          }
        }
      }
    },
...
  }
}